requirements.txt
pygame~=2.0.1
python-dotenv~=0.15.0
spotipy~=2.17.1
pyaudio>=0.2.11
SpeechRecognition>=3.8.1